Create constant file
