+++ app/src/main/java/com/example/news_app/network/RetrofitClient.kt
@@ -1,5 +1,6 @@
 package com.example.news_app.network
 
+import com.example.news_app.utils.Constants
 import com.google.gson.Gson
 import com.google.gson.GsonBuilder
 import dagger.Module
@@ -30,7 +31,7 @@ object RetrofitClient {
     @Singleton
     fun provideRetrofit(okHttpClient: OkHttpClient, gson: Gson): Retrofit {
         return Retrofit.Builder()
-            .baseUrl("https://jsonplaceholder.typicode.com/")
+            .baseUrl(Constants.BASE_URL)
             .client(okHttpClient)
             .addConverterFactory(GsonConverterFactory.create(gson))
             .build()
