+++ app/src/main/java/com/example/news_app/utils/Constants.kt
@@ -1,0 +1,6 @@
+package com.example.news_app.utils
+
+object Constants {
+    const val BASE_URL = "https://jsonplaceholder.typicode.com/"
+    const val IMAGE_URL = "https://techcrunch.com/wp-content/uploads/2024/05/Minecraft-keyart.jpg?resize=1200,720"
+}
